lib/errors: add interfaces for typed errors
